# Improve ORC reader performance