--- a/presto-orc/src/main/java/io/prestosql/orc/CachingOrcDataSource.java
+++ b/presto-orc/src/main/java/io/prestosql/orc/CachingOrcDataSource.java
@@ -17,6 +17,8 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
+import io.prestosql.orc.stream.MemoryOrcDataReader;
+import io.prestosql.orc.stream.OrcDataReader;
 
 import java.io.IOException;
 import java.util.Map;
@@ -93,17 +95,17 @@ public class CachingOrcDataSource
     }
 
     @Override
-    public <K> Map<K, OrcDataSourceInput> readFully(Map<K, DiskRange> diskRanges)
+    public <K> Map<K, OrcDataReader> readFully(Map<K, DiskRange> diskRanges)
             throws IOException
     {
-        ImmutableMap.Builder<K, OrcDataSourceInput> builder = ImmutableMap.builder();
+        ImmutableMap.Builder<K, OrcDataReader> builder = ImmutableMap.builder();
 
         // Assumption here: all disk ranges are in the same region. Therefore, serving them in arbitrary order
         // will not result in eviction of cache that otherwise could have served any of the DiskRanges provided.
         for (Map.Entry<K, DiskRange> entry : diskRanges.entrySet()) {
             DiskRange diskRange = entry.getValue();
             Slice buffer = readFully(diskRange.getOffset(), diskRange.getLength());
-            builder.put(entry.getKey(), new OrcDataSourceInput(buffer.getInput(), buffer.length()));
+            builder.put(entry.getKey(), new MemoryOrcDataReader(dataSource.getId(), buffer, buffer.length()));
         }
         return builder.build();
     }