--- a/presto-orc/src/main/java/io/prestosql/orc/OrcDataSource.java
+++ b/presto-orc/src/main/java/io/prestosql/orc/OrcDataSource.java
@@ -14,6 +14,7 @@
 package io.prestosql.orc;
 
 import io.airlift.slice.Slice;
+import io.prestosql.orc.stream.OrcDataReader;
 
 import java.io.Closeable;
 import java.io.IOException;
@@ -33,7 +34,7 @@ public interface OrcDataSource
     Slice readFully(long position, int length)
             throws IOException;
 
-    <K> Map<K, OrcDataSourceInput> readFully(Map<K, DiskRange> diskRanges)
+    <K> Map<K, OrcDataReader> readFully(Map<K, DiskRange> diskRanges)
             throws IOException;
 
     @Override