For TCP-DS, this saves about 9.5% of total CPU when running over gzip-compressed data.

Benchmarks are contained in `BenchmarkStreamReaders`. For this test, I used 3 forks each with
30 warm and 20 test iterations.  The results are average nanoseconds per value read.

Benchmark method  | Old |    | Error| New  |   | Error| Speedup
:-----------------|-----:|---|-----:|-----:|---|-----:|-----:
readBooleanNoNull | 2.48 | ± | 0.04 | 0.67 | ± | 0.01 | 3.73
readBooleanWithNull | 9.49 | ± | 0.27 | 2.01 | ± | 0.02 | 4.72
readByteNoNull | 2.83 | ± | 0.05 | 0.45 | ± | 0.02 | 6.23
readByteWithNull | 8.60 | ± | 0.30 | 1.95 | ± | 0.02 | 4.41
readDoubleNoNull | 15.61 | ± | 0.17 | 1.31 | ±|  0.07|  11.89
readDoubleWithNull | 16.14 | ± | 0.25 | 2.83 | ± | 0.04 | 5.70
readFloatNoNull | 14.64 | ± | 0.22 | 0.81 | ±|  0.04|  18.11
readFloatWithNull | 15.77 | ± | 0.22 | 2.08 | ± | 0.04 | 7.60
readIntNoNull | 7.83 | ± | 0.08 | 2.56 | ± | 0.03 | 3.06
readIntWithNull | 13.62 | ± | 0.18 | 2.99 | ± | 0.05 | 4.55
readLongDecimalNoNull | 82.88 | ± | 0.73 | 16.85 | ± | 0.27 | 4.92
readLongDecimalWithNull | 55.30 | ± | 0.81 | 11.77 | ± | 0.19 | 4.70
readLongNoNull | 9.43 | ± | 0.19 | 3.33 | ± | 0.06 | 2.83
readLongWithNull | 14.13 | ± | 0.17 | 3.26 | ± | 0.06 | 4.34
readShortDecimalNoNull | 27.35 | ± | 0.51 | 9.62 | ± | 0.15 | 2.84
readShortDecimalWithNull | 20.81 | ± | 0.30 | 6.96 | ± | 0.11 | 2.99
readShortNoNull | 6.79 | ± | 0.11 | 2.55 | ± | 0.03 | 2.66
readShortWithNull | 11.19 | ± | 0.12 | 2.95 | ± | 0.04 | 3.79
readSliceDictionaryNoNull | 3.31 | ± | 0.04 | 2.22 | ± | 0.02 | 1.49
readSliceDictionaryWithNull | 8.65 | ± | 0.10 | 3.67 | ± | 0.05 | 2.36
readSliceDirectNoNull | 34.46 | ± | 0.47 | 3.94 | ± | 0.08 | 8.76
readSliceDirectWithNull | 33.57 | ± | 0.56 | 6.99 | ± | 0.11 | 4.80
readTimestampNoNull | 17.91 | ± | 0.30 | 13.55 | ± | 0.20 | 1.32
readTimestampWithNull | 16.07 | ± | 0.24 | 11.77 | ± | 0.12 | 1.36