@sopel39 2019-03-28
Here are the actual benchmark results (both wall time and CPU): https://s3.us-east-2.amazonaws.com/starburstdata/karol/Benchmarks+comparison-orc_improvements.pdf